Added optional integration for AWS S3
The credentials and settings are passed in via environment variables.
The gem will look in the cwd for a .env file, and use that if present.
See .env.sample and the README for more info. Files will be
automatically added whenever the release command is run, using the same
structure they have in the release folder