Add paperback check command to verify dependencies
Create a quick script to ensure that dependencies have been satisfied on
the Docker image (and anywhere else, really, since this could still be
run locally). Add a spec that runs and verifies this by building the
image.

This lays some groundwork for Dockerfile refactors I have in mind to
support a couple of use-cases:

* Insert a new image that only handles Pandoc/LaTeX, then base a
  Paperback image on that instead of directly on ruby-2-2
* Don't use ruby-2-2 as it's a bloated image providing more than we
  really need. RVM-based images worked well for Scriptorium's Softcover
  image.

A couple of semi-related changes snuck in here that could possibly be
split into separate commits:

* Change entrypoint to a quick bash script that checks whether the
  command is a subcommand of Paperback and `exec`s otherwise (this is a
  standard practice in writing Dockerfiles when the command is
  unrecognized)
* Switch to tabs in Dockerfile (since it's shell, this makes more sense
  to me than spaces)
* Minor change in apt-get command to join update/install lines per
  [Dockerfile Best-Practices](https://docs.docker.com/articles/dockerfile_best-practices/#sort-multi-line-arguments)
  example.