Allow customization of page size
Previously Paperback was hardcoding the book geometry in command line arguments
sent to pandoc. Even though the pdf template supported custom page sizes, the
ones passed in via the command line would always override the sizes specified in
your book.

This changes it so the default size of 6x9 is only used if other geometry is not
provided in your own book.

To provide the geometry in your book, in the yaml frontmatter, put:

```
geometry: paperwidth=4in, paperheight=6in
```