Introduce a website format, used on thoughtbot.com
This introduces a new format, `Website` which uses pandoc's ChunkedHTML
functionality, with a custom template to generate Erb partials that can
then be used by thoughtbot.com.

There was previously no way to have a format which was different than
the extension/built-in types of pandoc, so this introduces an optional
separation between the name of the format in Paperback and the pandoc
format.