Don't link top level headers
Currently, all headers made into links in the website build, but with
the split-level set to 2, this means that all h1s and h2s, which are the
first headers on the page are also made into links, which is awkward.

This changes it so that only headers within the content (greater than
level 2) are made into links.